feat(feature): update snack bar duration - 현재 스낵바가 표기되는 시간이 예상보다 길어 표시되는 기간을 적절하게 수정합니다.

+++ feature/login/reset-password/src/main/kotlin/com/gradation/lift/feature/login/resetPassword/navigation/ResetPasswordRoute.kt
@@ -47,7 +47,7 @@ fun ResetPasswordRoute(
         is UpdatePasswordState.Fail -> {
             LaunchedEffect(updatePasswordStateResult.message) {
                 resetPasswordScreenState.snackbarHostState.showSnackbar(
-                    message = updatePasswordStateResult.message, duration = SnackbarDuration.Long
+                    message = updatePasswordStateResult.message, duration = SnackbarDuration.Short
                 )
                 updateUpdatePasswordState(UpdatePasswordState.None)
             }
